+++ src/components/UserPanel/UserSwitch/UserSwitch.tsx
@@ -1,4 +1,5 @@
 import React, { useState } from "react";
+
 import { useUserContext } from "../../../pages/Administration/user/UserProviderContext";
 
 const AUTHENTICATED_USER_ID = "selectedUserId";
