+++ src/contexts/LanguageProviderContext.tsx
@@ -19,7 +19,7 @@ interface LanguageContextProps {
 }
 
 const LanguageContext = createContext<LanguageContextProps | undefined>(
-  undefined
+  undefined,
 );
 
 const LANGUAGE_STORAGE_KEY = "selectedLanguage";
@@ -41,14 +41,11 @@ export const LanguageProvider = ({ children }: PropsWithChildren) => {
       selectedLanguage,
       setSelectedLanguage: (newLanguage: Language) => {
         setSelectedLanguage(newLanguage);
-        localStorage.setItem(
-          LANGUAGE_STORAGE_KEY,
-          JSON.stringify(newLanguage)
-        );
+        localStorage.setItem(LANGUAGE_STORAGE_KEY, JSON.stringify(newLanguage));
       },
       t: (key) => translations[selectedLanguage][key] || key,
     }),
-    [selectedLanguage]
+    [selectedLanguage],
   );
 
   return (
