+++ src/pages/Course/CourseManagement/course/constants.tsx
@@ -1,0 +1,6 @@
+export const departmentOptions = [
+  { label: "All", value: "all" },
+  { label: "Java", value: "java" },
+  { label: ".NET", value: ".net" },
+  { label: "SAP", value: "sap" },
+];
