+++ src/pages/Course/CourseManagement/course/types.ts
@@ -1,0 +1,5 @@
+export interface FormErrors {
+  courseName: string;
+  numberOfParticipants: string;
+  participantGroups: string;
+}
