+++ src/pages/Course/CourseOverview/CourseOverview.css
@@ -1,0 +1,68 @@
+.course-overview {
+  width: 100%;
+  box-sizing: border-box;
+  padding: 6.25rem 1.5rem 3rem;
+  background-color: whitesmoke;
+  min-height: 100vh;
+  display: flex;
+  flex-direction: column;
+}
+
+.course-overview-title {
+  color: #333;
+  font-weight: 500;
+  font-size: 1.75rem;
+}
+
+.toggle-container {
+  display: flex;
+  align-items: center;
+  margin-bottom: 1.5rem;
+  gap: 0.2rem;
+}
+
+.toggle-button {
+  background: none;
+  border: none;
+  cursor: pointer;
+}
+
+.toggle-switch {
+  position: relative;
+  width: 2.5rem;
+  height: 1.25rem;
+}
+
+.toggle-track {
+  width: 100%;
+  height: 100%;
+  border-radius: 624rem;
+  background-color: #d1d5db;
+  transition: background-color 0.2s ease-in-out;
+}
+
+.toggle-track.active {
+  background-color: #92b9f7;
+}
+
+.toggle-thumb {
+  position: absolute;
+  top: 0.125rem;
+  left: 0.125rem;
+  width: 1rem;
+  height: 1rem;
+  background-color: white;
+  border-radius: 50%;
+  transition: transform 0.2s ease-in-out;
+}
+
+.toggle-thumb.active {
+  transform: translateX(1.25rem);
+  background-color: #0062ff;
+}
+
+.toggle-label {
+  color: #374151;
+  font-weight: 600;
+  font-size: 0.87rem;
+}
