+++ src/pages/Course/CourseOverview/course/components/OpenCourseBtn.tsx
@@ -1,0 +1,11 @@
+import { IoMdSearch } from "react-icons/io";
+
+const OpenCourseBtn = () => {
+  return (
+    <button className="action-btn">
+      <IoMdSearch size={20} />
+    </button>
+  );
+};
+
+export default OpenCourseBtn;
