+++ src/pages/Course/constants.tsx
@@ -1,0 +1,20 @@
+export const defaultCourseData = [
+  {
+    id: "a1f4c5d6-1234-4b7c-8a2e-91f2ef1c3a01",
+    name: "Java Basics",
+    department: "java",
+    classification: "technical",
+    participantGroups: ["developers"],
+    numberOfParticipants: 25,
+  },
+  {
+    id: "b2e6d7a8-2345-4c8d-9b3f-a2e3f4d5b6c2",
+    name: ".NET Advanced",
+    department: ".net",
+    classification: "technical",
+    participantGroups: ["developers"],
+    numberOfParticipants: 30,
+  },
+];
+
+export const COURSES_STORAGE_KEY = "courses";
