+++ src/routes/index.tsx
@@ -2,9 +2,11 @@ import { useState } from "react";
 import { createBrowserRouter, Navigate, Outlet } from "react-router-dom";
 
 import { Navbar, Sidebar } from "../components";
+import CourseLayout from "./CourseLayout";
 import Administration from "../pages/Administration/Administration";
+import CourseManagement from "../pages/Course/CourseManagement/CourseManagement";
+import CourseOverview from "../pages/Course/CourseOverview/CourseOverview";
 import Home from "../pages/Home/Home";
-import CourseManagement from "../pages/CourseManagement/CourseManagement";
 
 const Layout = () => {
   const [sidebarOpen, setSidebarOpen] = useState<boolean>(false);
@@ -42,8 +44,18 @@ const router = createBrowserRouter([
         element: <Administration />,
       },
       {
-        path: "/course_management",
-        element: <CourseManagement />,
+        path: "/course",
+        element: <CourseLayout />,
+        children: [
+          {
+            path: "management",
+            element: <CourseManagement />,
+          },
+          {
+            path: "overview",
+            element: <CourseOverview />,
+          },
+        ],
       },
     ],
   },
