+++ src/utils/utils.tsx
@@ -25,14 +25,14 @@ export const Views: ViewTypes[] = [
   {
     id: "1",
     title: "courseManagement",
-    path: "/course_management",
+    path: "/course/management",
     image:
       "https://media.istockphoto.com/id/1011792694/photo/elevated-view-of-a-busy-open-plan-office.jpg?s=612x612&w=0&k=20&c=f-bXShoO_CyU0f1uSIDMk1CXj2IgCRqkG3KKOgBua9o=",
   },
   {
     id: "2",
     title: "courseOverview",
-    path: "/course_overview",
+    path: "/course/overview",
     image: "https://images.unsplash.com/photo-1504384308090-c894fdcc538d",
   },
   {
